[ refactor ] cardLists 패칭 로직 분리 :: useCardLists
