+++ src/components/CardCollectionPage/index.tsx
@@ -1,7 +1,7 @@
-import { useState } from "react";
 import { useLocation } from "react-router-dom";
 
-import { CardList, CardsTypeLocation } from "../../types/cardCollection";
+import useCardLists from "../../core/api/cardCollection";
+import { CardsTypeLocation } from "../../types/cardCollection";
 import Header from "../common/Header";
 import useModal from "../common/hooks/useModal";
 import useScroll from "../common/hooks/useScroll";
@@ -15,8 +15,7 @@ export default function CardCollectionPage() {
 
   const location = useLocation();
   const cardsTypeLoaction = location.state as CardsTypeLocation;
-
-  const [cardLists, setCardLists] = useState<CardList[] | null>(null);
+  const { cardLists, fetchCardListsWithFilter } = useCardLists(cardsTypeLoaction);
 
   const { isModalOpen: isFilterModalOpen, toggleModal: toggleFilterModal } = useModal();
   const { isModalOpen: isLoginModalOpen, toggleModal: toggleLoginModal } = useModal();
@@ -29,11 +28,12 @@ export default function CardCollectionPage() {
         openLoginModalHandler={toggleLoginModal}
         cardsTypeLoaction={cardsTypeLoaction}
         cardLists={cardLists}
-        setCardLists={setCardLists}
       />
 
       {isLoginModalOpen && <LoginModal closeHandler={toggleLoginModal} contents={"북마크기능인 마이피클을"} />}
-      {isFilterModalOpen && <FilterModal closeHandler={toggleFilterModal} setCardLists={setCardLists} />}
+      {isFilterModalOpen && (
+        <FilterModal closeHandler={toggleFilterModal} fetchCardListsWithFilter={fetchCardListsWithFilter} />
+      )}
     </St.MainPage>
   );
 }
